refactor(UserPage): rename Widgets => widgets (#409)
