+++ src/client/components/UserPage/CreateUrlModal/CreateLinkForm.jsx
@@ -19,12 +19,12 @@ import {
 } from '../../../../shared/util/validation'
 import ModalMargins from './ModalMargins'
 import refreshIcon from './assets/refresh-icon.svg'
-import LinkIcon from '../Widgets/LinkIcon'
-import FileIcon from '../Widgets/FileIcon'
+import LinkIcon from '../widgets/LinkIcon'
+import FileIcon from '../widgets/FileIcon'
 import { formatBytes } from '../../../util/format'
 import CollapsibleMessage from '../../CollapsibleMessage'
 import { CollapsibleMessageType } from '../../CollapsibleMessage/types'
-import FileInputField from '../Widgets/FileInputField'
+import FileInputField from '../widgets/FileInputField'
 
 // Height of the text field in the create link dialog.
 const TEXT_FIELD_HEIGHT = 44
