+++ src/client/components/UserPage/widgets/CopyButton.tsx
