+++ src/client/components/UserPage/widgets/LinkIcon.tsx
