+++ src/client/components/UserPage/widgets/OnClickTooltip.tsx
